refactor(ooda_actions/advance_goal): extract tests to sibling (#1266)
- advance_goal.rs: 900 -> 659 LOC
- new tests_advance_goal.rs (241 LOC) merging mod tests + mod inflight_tests
- widen validate_subordinate_completion to pub(super)
